header, Reservation 분리
