+++ .idea/jsLinters/eslint.xml
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="EslintConfiguration">
+    <work-dir-patterns value="$PROJECT_DIR$/app-web" />
     <option name="fix-on-save" value="true" />
   </component>
 </project>
