+++ app-web/package.json
@@ -19,7 +19,8 @@
     "react": "^18.0.0",
     "react-dom": "^18.0.0",
     "react-hook-form": "^7.33.1",
-    "react-query": "^3.39.1"
+    "react-query": "^3.39.1",
+    "vite-plugin-babel": "^1.1.0"
   },
   "devDependencies": {
     "@babel/core": "^7.18.6",
