+++ app-web/src/Header.tsx
@@ -1,0 +1,9 @@
+function Header() {
+    return (
+        <header>
+            <h1>코드숨 공부방 좌석 예약 앱</h1>
+        </header>
+    );
+}
+
+export default Header;
