+++ app-web/vite.config.ts
@@ -1,8 +1,12 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
+import babel from 'vite-plugin-babel';
 
 // https://vitejs.dev/config/
 export default defineConfig({
   base: '/my-seat/',
-  plugins: [react()]
+  plugins: [
+      react(),
+      babel(),
+  ]
 })
